Renamed dpos utils and updated imports
