+++ packages/block/store/action.js
@@ -1,7 +1,7 @@
 import { ROUND_LENGTH } from '@dpos/validator/constants';
 import { convertUnixSecondsToLiskEpochSeconds } from '@views/utilities/dateTime';
 import { getBlocks } from '@block/utilities/api';
-import { getForgers, getDelegates } from '@dpos/validator/utiles/api';
+import { getForgers, getDelegates } from '@dpos/validator/utils/api';
 import actionTypes from './actionTypes';
 
 /**
